+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -15,15 +15,15 @@ import org.apache.coyote.Processor;
 import org.apache.coyote.http11.cookie.HttpCookie;
 import org.apache.coyote.http11.cookie.SessionCookieFactory;
 import org.apache.coyote.http11.exception.HttpStatusException;
-import org.apache.coyote.http11.httprequest.HttpMethod;
-import org.apache.coyote.http11.httprequest.HttpRequest;
-import org.apache.coyote.http11.httpresponse.HttpResponse;
-import org.apache.coyote.http11.httpresponse.HttpStatusCode;
-import org.apache.coyote.http11.parser.FormUrlEncodedHttpRequestParser;
-import org.apache.coyote.http11.parser.HttpResponseParser;
-import org.apache.coyote.http11.parser.SessionParser;
+import org.apache.coyote.http11.request.FormUrlEncodedHttpRequestParser;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.startline.HttpMethod;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.HttpResponseParser;
+import org.apache.coyote.http11.response.startline.HttpStatusCode;
 import org.apache.coyote.http11.session.Session;
 import org.apache.coyote.http11.session.SessionManager;
+import org.apache.coyote.http11.session.SessionParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
