+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestParser.java
@@ -1,12 +1,12 @@
-package org.apache.coyote.http11.parser;
+package org.apache.coyote.http11.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import org.apache.coyote.http11.httprequest.RequestBody;
-import org.apache.coyote.http11.httprequest.RequestHeaders;
-import org.apache.coyote.http11.httprequest.RequestLine;
+import org.apache.coyote.http11.request.body.RequestBody;
+import org.apache.coyote.http11.request.header.RequestHeaders;
+import org.apache.coyote.http11.request.startline.RequestLine;
 
 public abstract class HttpRequestParser {
