+++ tomcat/src/main/java/org/apache/coyote/http11/request/startline/RequestUri.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.httprequest;
+package org.apache.coyote.http11.request.startline;
 
 import java.util.HashMap;
 import java.util.Map;
