+++ tomcat/src/main/java/org/apache/coyote/http11/response/body/ResponseBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.httpresponse;
+package org.apache.coyote.http11.response.body;
 
 import java.nio.charset.StandardCharsets;
