+++ tomcat/src/main/java/org/apache/coyote/http11/response/startline/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.httpresponse;
+package org.apache.coyote.http11.response.startline;
 
 public enum HttpStatusCode {
